util/event: Infrastructure for the virtualization of spdk_thread.
This has the base infrastructure set up for the
virtualization of spdk_thread.

The design is to embed pollers/timers/event rings
in the spdk_thread infrastructure. The native threads
(i.e. DPDK threads) use the reactor polling loop. When this
functionality is enabled, the reactors contain the virtual
thread polling functions. No other pollers run on the reactor.

This allows an app to run a virtual thread on any reactor
and move it around. It also allows an app to create new
virtual threads (depending on load/ports) and "assign" it to
any reactor.

The main changes here are:
1. Create a new TLS to identify the virtual thread.
2. Additions to the spdk_thread data structure.
3. Accessor functions around the spdk_thread data structure.
4. UT code changes for the util library changes.
5. New UT suite to test the dynamic threading.
6. Introduced a flag for dynamic threading in the config file.
7. Reactor/APP changes to use new mechanism.
8. Fixed poll groups for NVMF.
9. Updated CIT for NVMf to use dynamic threading flag.

All CIT's passed in the Netapp environment with and without the flag.
[We are aware that we don't run all the CITs that Intel does]

Next steps:
1. Other apps need to be modified  to work with this logic.
2. Remove switch once approved, thus making code flow simpler.
3. Add logic in apps that show how threads can be moved between
reactors (possibly need a thread state to allow threads to break
from the polling loop).

Known Issues:
1. During startup, there exists a race condition where the
threads have not completed initialization before the app tries
adding events to the event ring.
2. Currently threads will run on only one reactor. The infrastructure
to break a thread from a reactor has not been added. (This is part
of "Next steps").

Updates based on comments and CIT results

1. Code review comments from GerritHub addressed here.
Note: spdk_event_allocate is NOT called in the data path.

2. Fixed Freebsd compilation issues (caught in the Intel CIT).
The failure is due to pthread_t being a pointer in Freebsd.

Change-Id: I6d878950954c0f4d345bbe91f7037c3742b93284
Signed-off-by: John Meneghini <johnm@netapp.com>
Signed-off-by: Madhu Pai <mpai@netapp.com>