scsi: Unify the name of and refactor functions about hot-remove
Change the name of hot plug to hot remove not to confuse.

Poller of hot remove runs with 10 usec period, and so unconditional
invocation of poller will be allowed.

Hot remove is not critical for performance, and so unconditional
message sending will be allowed.

it is unexpected that SCSI device is empty when deleted, and so
take error log.

Change-Id: Ifc8de3452de72a08eba0e9ea942217e5342b2ef4
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>