Update patch set 4
Patch Set 4:

I think this patch can be broken up into a series of self-contained commits, each of which would pass the full test suite. That will make this much more manageable. I've started the process by breaking the threading abstraction out of util and into a dedicated library in lib/thread and rebased this patch on top of that. However, there's a lot more to do still.

I think the series should look something like this:

1) Move poller list and event ring into spdk_thread. Add a function to poll the thread. The reactor should create one thread and poll it in a while loop, but otherwise maintain the 1:1 mapping here.

2) Add a thread-local variable to lib/thread that sets the "current" thread - add an API to set the current thread, make spdk_get_thread use the thread-local variable for look up, and set the current thread on reactor start.

3) Add a ring of threads to the reactor. The reactor can still create one thread at the beginning of time and place it on the ring.

4) Add functions to "schedule" threads onto reactors and use that to bootstrap the application framework as opposed to each reactor automatically constructing a thread when it begins running.

Patch-set: 4