Update patch set 1
Patch Set 1:

Note that vector commands are optional. Some vendors are thinking of only doing LBA Read/Write/DSM for the commands.

There is missing a bit in the spec to detect this. I'm working on a 2.0a revision that'll add it in the MCCAP field as bit 2 (should had been bit 0).

Patch-set: 1