Update patch set 2
Patch Set 2:

This is a note to say that this patch failed verification because of
a problem with the asytle utility on the ubuntu 16 CIT testbed.  I have seen this problem before, and I have reproduced this problem on Ubuntu 16.  Not that the same patch does NOT display this problem on any other
CIT testbed, so this is a CIT tool and configuration problem, not a problem with the patch.

See: https://github.com/spdk/spdk/issues/334

Patch-set: 2
Reviewer: John Meneghini <1011271@d5d70762-12d0-45a1-890d-524b12d3f735>