shared_lib: only export spdk_* symbols
Add a basic linker version script to limit exported symbols to those
intended to be public API.  Currently, many internal-only functions are
also named spdk_*, but this will at least limit exported symbols to
those that will not pollute the global namespace.

This also does not assign a version to any symbols; in the future, we
may want to extend the version script to enable ABI compatibility, but
for now, the version script is just used to control symbol visibility.

Change-Id: I1f020300b3288a026b47e9b1466ebc72952d9401
Signed-off-by: Daniel Verkamp <daniel.verkamp@intel.com>
Reviewed-on: https://review.gerrithub.io/415897
Tested-by: SPDK Automated Test System <sys_sgsw@intel.com>
Reviewed-by: Ben Walker <benjamin.walker@intel.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>