util/event: Infrastructure for the virtualization of spdk_thread.
This has the base infrastructure set up for the
virtualization of spdk_thread.

The design is to embed pollers/timers/event rings
in the spdk_thread infrastructure. The native threads
(i.e. DPDK threads) use the reactor polling loop. When this
functionality is fully enabled, the reactors contain the virtual
thread polling functions. No other pollers run on the reactor.

This allows an app to run a virtual thread on any reactor
and move it around. It also allows an app to create new
virtual threads (depending on load/ports) and "assign" it to
any reactor.

Implementation of the above functionality is broken up in a
series of patches.

This patch specifically has:
1. Additions to the spdk_thread data structure for pollers, timers
   and events.
2. Accessor functions around the spdk_thread data structure.
3. UT code changes for the util library changes.
4. Update to the reactor polling logic to route events, pollers
   and timers to the corresponding thread.
5. A 1-1 mapping between a reactor and a spdk_thread.

Next steps:
1. Remove the dependency (1-1 mapping) between the thread
   and the reactor.
2. Add a TLS value that identifies the virtual thread.

Change-Id: I6d878950954c0f4d345bbe91f7037c3742b93284
Signed-off-by: John Meneghini <johnm@netapp.com>
Signed-off-by: Madhu Pai <mpai@netapp.com>