env: create secondary own mempool if same name existing
This patch added the logic to handle the case where named
mempool already allocated by the primary process and failed
the startup of the secondary process. Instead, create the
mempool with the unique name from the secondary process.

For the destroy of the allocated mempool, it will be same
to rely on each process's cleanup.

Change-Id: I797c28d18afbfd5ec4e8b16f4e815a7b1f8ecc54
Signed-off-by: GangCao <gang.cao@intel.com>