scsi: Don't free IO channel when it is not allocated at LUN removal
Handling IO channel of LUN is not clear in iSCSI. Current code frees
IO channel even if it is not allocated.

Besides, when IO channel is not allocated, there is no outstanding
iSCSI tasks and hence check of outstanding iSCSI tasks is not necessary
in this case.

LUN IO channel management in iSCSI still have issues to be fixed.
This is a small step to go forward.

Change-Id: Id3721180422364dfc4d9309f3a22ce0a3f766f82
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>