Update patch set 3
Patch Set 3: Code-Review-1

I've added a bunch or people from Broadcom and NetApp to this review.  As I stated in the email this morning, I think we need to align around what the over-all dev-io lifecycle is here.  This patch is fine in and of itself, but I don't want to integrate this into the master branch until we've agreed about some of the larger design issues are with the bdevio life cycle.

Patch-set: 3
Reviewer: John Meneghini <1011271@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1