go: empty Go package
The only purpose at the moment is to allow vendoring of the SPDK
source code into other Go projects which manage dependencies with
the "dep" tool. That tool is designed for Go projects and as a sanity
check rejects any dependency which does not have at least one Go file.

Change-Id: I22458d0895729ad7d8ea53e7541e9089da1d448a