iscsi initiator: comment out iscsi_destroy_context
When linked with some newest version in libiscsi, which
causes the segement. During the debug, it clearly points
out this. Since currently, we may not link with delegated
libiscsi, so the temp solution is to comment out this.
And we should also have a patch for libiscsi later.

This patch should solve the issue reported in github:

https://github.com/spdk/spdk/issues/340

Change-Id: Ib732cb2eef58ece585af9b8b9c186f414cb18ce0
Signed-off-by: Ziye Yang <optimistyzy@gmail.com>