thread: Make the thread internally manage events and pollers
The user of the thread library is now only responsible for
periodically calling spdk_thread_poll. Events (messages) and
pollers are handled internally.

In order to avoid changing all of the unit tests, the ability
to provide function pointers to change the behavior of
the message passing and poller registration is still in
the code. This should only be used from tests until they
are all converted.

Change-Id: Ie2c00ce1d57ca3710ed2c469cd711924768e23ef
Signed-off-by: Ben Walker <benjamin.walker@intel.com>