Update patch set 6
Patch Set 6: Code-Review-1

Instead of using memzones here, let's instead use spdk_dma_zmalloc with an alignment restriction.

If we pick an alignment that is bigger than or equal to the buffer size, and is evenly divisible by 2MB, then we are guaranteed that the buffer will not span a 2MB boundary and hence will be physically contiguous.

This is way simpler than using the memzone API.  I talked with Ben and Daniel and they agree on using this new approach as well.

Patch-set: 6
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1