vbdev/lvol: fix the error handling path
Issue: The base bdev is not closed.

When the bdev lvol examine the device,
if it is called in line 362 in _spdk_lvs_load_cb,
finally _vbdev_lvs_examine_cb is called, then we do
not do any work to destroy the bs_dev.

Also: We need to destroy the relatedd I/O channel,
however currently we did not have function in
to call since all the followings are based on
(struct spdk_blob_store *bs), so it is a
problem, which will cause the I/O channel leak.

Change-Id: I8459706d2d776b68874351b3b595cbf236639959
Signed-off-by: Ziye Yang <optimistyzy@gmail.com>