thread: don't immediately remove channel from list when put
When spdk_put_io_channel is called, if its the last reference,
we defer actual destruction of the channel, so that code
in the same context which may be referring to the channel
doesn't crash.

But it is possible that an io_channel for that same io_device
could be requested before the deferred message is processed.
This would result in a second io_channel being created for
that device on the same thread.

To avoid this case, don't remove the channel from the list when
ref reaches 0.  When the deferred message is processed, if the
count no longer equals 0, don't destroy the channel.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: Idb8d4705fda0eb9c338e4960430e04edbe537e05