Update patch set 1
Patch Set 1:

This is to fix the Github issue https://github.com/spdk/spdk/issues/353 where I have debugged that the I/O channel is never got freed so that in the spdk_scsi_lun_check_io_channel() call it is hanging and the bdev is not unregistered. 

The PRC call of delete_bdev also timed out at last.

Seems this issue is relating to below change.
https://github.com/spdk/spdk/commit/1e1e0dd70ce05a43e17b184e8e2dc6850d4c20fc#diff-50d1a1d47c736b4130caa87389386e0e

Could Shuhei give more information here what the original issue tried to fix? As from the code navigation, the I/O channel looks like only freed when the connection destroys and here it is just a RPC call to delete the bdev and trigger the LUN hot removal which seems like not relating to the connection destruction so that the I/O channel always there and hangs the delete bdev operation.

My change here may be not correct (verified it can fix the reported issue 353) and just to start the discussion on this patch to get this issue (Github 353) fixed properly.

Please Shuhei comment here.

Thanks,
Gang

Patch-set: 1
Reviewer: GangCao <1011207@d5d70762-12d0-45a1-890d-524b12d3f735>