Update patch set 3
Patch Set 3:

It's also quite difficult to do a serious review of this header file without the associated implementation. So while this file can get approved and merged now, it's at the risk of major refactor later on when the details of the implementation are revealed. I just want to make sure everyone is on the same page about what may happen down the road so there are no surprises.

In an ideal world, the code would not be broken up by file, but rather by logical groupings of functionality that can be reviewed as a unit. I understand that you may not have a nice series of patches that implement one thing at a time at this point, so you'd have to go back and break it all up manually which would be very time consuming. I'm not asking you to go through and do that for this patch, but for newly written code its much less risky to break things into logical units (as opposed to file units).

Patch-set: 3
Reviewer: Ben Walker <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>