Update patch set 4
Patch Set 4:

(1 comment)

In our bdev this API would have to be called regardless of whether the IO failed or not.  The rules about how and IOVEC is run down after a READ or WRITE completes are BDEV specific.  My problem here is that I need an BDEV API that preserves the BDEV IO type, and allows my BDEV to replace this code with my own.

Patch-set: 4
CC: John Meneghini <1011271@d5d70762-12d0-45a1-890d-524b12d3f735>