Update patch set 4
Patch Set 4:

Hi John,

So when the zcopy_end is called, you need to know whether the original zcopy_start was done with populate = true (for a read) or false (for a write)?

If so - that seems reasonable.

-Jim

Patch-set: 4
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>