Update patch set 15
Patch Set 15: Code-Review+1

(7 comments)

There's still more work to do here but I'd like to merge this in ASAP.

The sooner we get people using this stuff the sooner we can fix the bugs and make improvements.

Also, there are a couple of issues that are not related to the code here. 1) Oracle now requires companies to purchase a license if they want to use Virtual Box Extensions, and 2) Vagrant requires a license if you want to use their code with VmWare Fusion.

I have this code running on VmWare Fusion now, but that work is not included in these patches.

Patch-set: 15
Reviewer: John Meneghini <1011271@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+1