Update patch set 1
Patch Set 1:

Hi Takeshi,

I am curious - did you use the SPDK rocksdb fork?  In the SPDK rocksdb fork, we modify the BGThreadWrapper to do an SpdkInitializeThread().

That is not ideal and your solution would remove the need for that part of the fork.  But if you are using it, then I am wondering which thread is making these calls to SpdkEnv?

Thanks,

-Jim

Patch-set: 1
CC: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>