Update patch set 39
Patch Set 39:

I'm fine with removing the lock, but keeping the spdk_lvol_deletable check.

Typically functions like this are problematic, because it may return true (meaning it can be deleted), but by time you act on the data, that state could have changed.  But for blobstore-related code like this, all of this must happen on the metadata thread so that doesn't really hold.

Patch-set: 39
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>