Update patch set 14
Patch Set 14:

> Patch Set 14:
> 
> (1 comment)

I understand that if you are in "in_submit_request" a msg is posted back for io_complete, however that would be a context switch. Today we can allocate buffers in the same thread and submit write request to bdev module. For each IO if we have to switch there would be a performance penalty. If we design to submit in the same thread back, then we will have the re-entrant problem.

Patch-set: 14
Reviewer: Sriram Popuri <1011327@d5d70762-12d0-45a1-890d-524b12d3f735>