Update patch set 14
Patch Set 14:

It's not a context switch, in the sense that the kernel scheduler can swap the thread out. It's the same threading processing the entire time for almost all implementations of zcopy. It just happens to place the request on a queue, unwind it's current stack, then grab the request from the queue again to avoid very deep stacks. This is a very fast and effective method of managing the stack and will not have any measurable performance impact.

Patch-set: 14
Reviewer: Ben Walker <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>