Update patch set 14
Patch Set 14: Code-Review-1

There is not a context switch in terms of the kernel scheduler. But, this an unnecessary reposting of an event into the ring. This comes with a performance penalty because posting and removing elements from the ring has a cost. 
This is especially true here because these are MP-SC queues. Multiple backend threads post callback events into the same ring. Although the rings are lockless, they are not waitless.
This reposting and de-queueing is avoided in the original zero copy patch that was proposed earlier.

Patch-set: 14
Reviewer: Madhu Pai <1014342@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1